protocol clean up (#185)
