+++ .github/workflows/merge.yaml
@@ -67,7 +67,7 @@ jobs:
         uses: actions-rs/cargo@v1
         with:
           command: test
-          args: --lib --all
+          args: --all
       - name: code-coverage
         run: cargo tarpaulin --exclude-files proof-of-concepts/ -o Lcov --output-dir ./coverage
       - name: coveralls
