+++ .github/workflows/pull-requests.yml
@@ -73,7 +73,7 @@ jobs:
         uses: actions-rs/cargo@v1
         with:
           command: test
-          args: --lib --all
+          args: --all
       - name: code-coverage
         if: matrix.os == 'ubuntu'
         run: cargo tarpaulin --exclude-files proof-of-concepts/ -o Lcov --output-dir ./coverage
