+++ Cargo.toml
@@ -6,5 +6,4 @@ members = [
     "src/sql_engine",
     "src/sql_types",
     "src/storage",
-    "src/test_helpers"
 ]
