--- src/test_helpers/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "test_helpers"
-version = "0.1.0"
-authors = ["Alex Dukhno <alex.dukhno@icloud.com>"]
-edition = "2018"
-publish = false
-
-[dependencies]
-async-std = { version = "1.6.2", features = ["attributes"] }
-tempfile = "3.1.0"
-bytes = "0.5"
-storage = { path = "../storage" }
-kernel = { path = "../kernel" }
-
-[dev-dependencies]
-rstest = "0.6.4"
